Auto complete Resp , mila ajustement

+++ src/main/java/afg/achat/afgApprovAchat/service/fournisseur/ResponsableFournisseurService.java
@@ -17,7 +17,7 @@ public class ResponsableFournisseurService {
                 .findFirstByFournisseur_IdOrderByDateAffectationDesc(idFournisseur);
     }
 
-    public Optional<ResponsableFournisseur> getById(Long id) {
+    public Optional<ResponsableFournisseur> getById(int id) {
         return responsableFournisseurRepo.findById(id);
     }
